Do not abort checking on proxies which return a Content-Length header.
Note that although returning a Content-Length header is invalid behavior
for a proxy, some proxies still return it and also succeed the CONNECT
request correctly. For this reason, we do not abort the check, and
instead rely on the -c method to accurately determine if it's a true
proxy or not.